Relicense to LGPL 2.1 or later
For the record, we main contributors (https://github.com/hardywood-k,
Hong Minhee, Kijun Seo, and Swen Mun) verbally discussed about relicensing
Libplanet together, and we decided to license it under LGPL 2.1 or later.
There are several rationales behind it, but the most influential one was
the VLC's case [1], which shares many considerations among us.

We want Libplanet and its derivations to be always open and transparent,
while also to be widely adopted to end-users at a time.  For the former
goal, copyleft licenses are preferable to permissive licenses, and for
the latter goal, open source licenses compatible to the terms of App Store
and Google Play.  It's basically the same to the VLC's case, so we
finally have the same decision: LGPL 2.1.

[1] https://www.videolan.org/press/lgpl-libvlc.html