Interface to store addresses mask into blocks
Note: Although I added an ad-hoc field separated from Block type,
this leads the order of storing multiple blocks to be forced
unless there are information about corresponding addresses masks
for these blocks to put in.  (A mask for a block needs to be
made on top of the mask for its previous block.)

For example, if we want to retrieve a large amount of blocks
out of order, we need to serialize the order of block insertions,
from the genesis block to the topmost block, or retrieve pairs
of block and its mask.  Even if we retrieve blocks along with
their masks, as masks are neither signed nor hashed with nonce,
we implement extra validations for masks.  It doesn't make only
dealing with blocks complex and painful, but also these jobs
bug-prone.

So I suggest to put the mask field into the Block type,
so that it is fully integrated into blockchain and all
things are validated along with, in a consistent way.