Revert prev workaround for LiteDB and make it flush instead
It turns out that LiteDB's data corruption
(https://github.com/mbdavid/LiteDB/issues/1268) seems to happen
due to sudden termination of program.  So the workaround made in the
previous patch (https://github.com/planetarium/libplanet/pull/386)
was reverted for the most part, and a new option named `flush` was
added to `LiteDBStore()` constructor.