Catch ArgumentOutOfRangeException; return null instead during block sync
As BlockLocator() constructor already is ready for indexBlockHash()
callback returning null (but is not ready for it throwing exceptions),
make it to catch unexpected ArgumentOutOfRangeException and return null
instead of throwing exceptions.

On the other hand, to investigate enough when such exception is thrown,
I left a log with local variable values too.

Closes https://github.com/planetarium/libplanet/issues/1346

[changelog skip]