Bump target framework to 6.0 for test projects and global.json
Notes:

When `Libplanet.Explorer` project is set to target `net6.0`, it
requires `6.x` version of `Microsoft.AspNetCore.Mvc.NewtonsoftJson`
package. However, since MSBuild bundled with Mono cannot correctly
interpret `net6.0` moniker, it is impossible to build the project
with mono. For this, a new target `ReleaseMono` was set up to
specifically exclude `Libplanet.Explorer*` projects from being
built with mono.

With `netcoreapp3.1`, only the executable produced with
`dotnet publish` was marked executable. However, with `net6.0`,
all the produced binaries are marked executable. Because of this,
`bin_name` had to be produced in another way, which at the moment
is looking whether it has `o+x` (or `g+x`) permission.