test(tx): Add max gas price and gas limit fields to Tx and TxInvoice
This commit adds `maxGasPrice` and `gasLimit` fields to `Tx` and `TxInvoice` classes respectively. These fields are used to limit the maximum gas price and gas used by a transaction. The values of these fields are set to `null` in the test cases because they are not relevant to these tests.

MEN002 (long lines) has been disabled for the test JSON data.